requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
pydantic>=2.5.0
networkx>=3.0
qrcode>=7.4
pillow>=10.0




